weak_link/requirements.txt
Django==4.2
import_export==0.3.1
numpy==1.24.3
pandas==1.5.3
Pillow==10.1.0
python-dotenv==1.0.0
scikit_learn==1.2.2
tablib==3.5.0
